% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyplot.r
\name{dyplot}
\alias{dyplot}
\title{Dynamic time series plots (wrapper to dygraphs)}
\usage{
dyplot(
  data,
  xcol = colnames(data)[1],
  ycol = setdiff(colnames(data), xcol),
  var_col = NULL,
  val_col = NULL,
  interval = NULL
)
}
\arguments{
\item{data}{a data.frame}

\item{xcol}{name of the column with the x-Axis variable. Ideally a datetime object}

\item{ycol}{names of the columns that are shown on the yaxis (used for a wide data fram)}

\item{var_col}{the column with the values for a data frame in long format}

\item{interval}{if you have missing rows you can specify the interval of you data, e.g. "day" or "year" or "hour" to fill the gaps}
}
\description{
The function is a wrapper to the dygraph function in the package dygraph. It shows
time series plots with interactive java script. While dygraph needs an xts time series
object, dyplot works with a data frame. Missing observations can be filled if period is
provided.
}
