% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyplot.r
\name{to.xts}
\alias{to.xts}
\title{Transform a data from to an xts object}
\usage{
to.xts(
  dat,
  time.col,
  interval = NULL,
  time = dat[[time.col]],
  fill = !is.null(interval)
)
}
\description{
Transform a data from to an xts object
}
